dpi resolution fixes

+++ ext/openpdf/src/main/java/net/sf/jasperreports/openpdf/producer/StandardPdfProducer.java
@@ -247,7 +247,7 @@ public class StandardPdfProducer implements PdfProducer
 	}
 	
 	@Override
-	public void setPageSize(PrintPageFormat pageFormat, int pageWidth, int pageHeight)
+	public void setPageSize(PrintPageFormat pageFormat, float pageWidth, float pageHeight)
 	{
 		Rectangle pageSize;
 		if (pageFormat.getOrientation() == OrientationEnum.LANDSCAPE)
@@ -433,7 +433,7 @@ public class StandardPdfProducer implements PdfProducer
 	}
 	
 	@Override
-	public PdfImage clipImage(PdfImage image, int clipWidth, int clipHeight, int translateX, int translateY) throws JRException
+	public PdfImage clipImage(PdfImage image, float clipWidth, float clipHeight, float translateX, float translateY) throws JRException
 	{
 		Image img = ((StandardImage)image).getImage();
 
@@ -615,7 +615,7 @@ public class StandardPdfProducer implements PdfProducer
 
 	@Override
 	public void addPdfPage(byte[] pdfData,
-			int x, int y, int width, int height)
+			float x, float y, float width, float height)
 	{
 		try (PdfReader pdfReader = new PdfReader(pdfData))
 		{
